Added: Global mouse hook.
Created two new files (Hooks.cs and MouseHook.cs) and moved general hook stuff into Hooks.cs.

First implantation of mouse hook, it is known that when the mouse hook is active and you click the minimize button, maximize button, and close button on the form it does lag for a moment and will lag if mouse is held down on those buttons.