Fixed FPSCurrent offset value
NOTE: When you get the property FPSLimit it
will not return the real value but 1000/real value.

Added a new function to the Client class,
SetFPSLimit(double value) this method will set
the FPSLimit to the given value (it will not convert).

Now if there are some other tibia client open
and the user select "New tibia client" it will
call the OpenMC method.

Fixed a little mistake in the packets Enums.cs.